requirements.txt
Django==4.2.0
psycopg2-binary==2.9.5
pytest==7.3.1
pytest-django==4.5.2
pytest-cov==4.1.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
python-dotenv==1.0.0
